TestCases
While testing I discovered you do not have to test for all the lengths of the edges. You can test each edge slope and if any are equal, you do not have a valid triangle. You can test for all vertices, and if any are in the same place then you do not have a valid triangle.